+++ RELEASE-NOTES.md
@@ -1,7 +1,11 @@
 cassdir release notes
 =====================
 
-0.1.0 - 2015-08-10
+0.1.0 - 2015-08-12
 ------------------
 
-- Preview release: `CassandraDirectory` implementation without locking.
+First release:
+
+- Cassandra-implementation of Lucene's Directory class.
+- Support caching.
+- Support fine-tune of Cassandra ConsistencyLevel per operation.